Move syncManager to engine
