+++ pkg/protocol/engine/syncmanager/trivialsyncmanager/syncmanager.go
@@ -10,7 +10,7 @@ import (
 	"github.com/iotaledger/iota-core/pkg/protocol/engine"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/blocks"
 	"github.com/iotaledger/iota-core/pkg/protocol/engine/clock"
-	"github.com/iotaledger/iota-core/pkg/protocol/syncmanager"
+	"github.com/iotaledger/iota-core/pkg/protocol/engine/syncmanager"
 	iotago "github.com/iotaledger/iota.go/v4"
 )
